+++ kraftwerk-core/src/main/java/fr/insee/kraftwerk/core/sequence/ControlInputSequence.java
@@ -7,9 +7,9 @@ import java.nio.file.Paths;
 import fr.insee.kraftwerk.core.Constants;
 import fr.insee.kraftwerk.core.exceptions.KraftwerkException;
 import fr.insee.kraftwerk.core.inputs.UserInputs;
-import lombok.extern.slf4j.Slf4j;
+import lombok.extern.log4j.Log4j2;
 
-@Slf4j
+@Log4j2
 public class ControlInputSequence {
 	
 	String defaultDirectory;