+++ kraftwerk-core/src/main/java/fr/insee/kraftwerk/core/utils/DateUtils.java
@@ -7,9 +7,9 @@ import java.util.Date;
 import java.util.TimeZone;
 import java.util.concurrent.TimeUnit;
 
-import lombok.extern.slf4j.Slf4j;
+import lombok.extern.log4j.Log4j2;
 
-@Slf4j
+@Log4j2
 public class DateUtils {
 
 	private DateUtils() {