+++ kraftwerk-core/src/main/java/fr/insee/kraftwerk/core/utils/TextFileReader.java
@@ -6,9 +6,9 @@ import java.io.IOException;
 import java.io.Reader;
 import java.nio.file.Path;
 
-import lombok.extern.slf4j.Slf4j;
+import lombok.extern.log4j.Log4j2;
 
-@Slf4j
+@Log4j2
 public class TextFileReader {